python-backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
redis
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
pydantic
